Allow certain cases in serialization using flexible
In certain cases the serialization and deserialization functions needs to be defined separately. Currently if the user defined “serialize” and “deserialize” functions, and wants to use “flexible” format (such as using ar.archive(data)) and in serialization function, if the input format is, e.x. function returned values (which shall be valid since no modification happens to the input), the “serialize” function will failed by static_assert error. The MR keeps the previous static_assertion requirement for “deserialize” function and allowed the above case to run.